+++ CHANGELOG.md
@@ -4,6 +4,9 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## [Unreleased]
+### Added
+- Add `Troubleshooting` page with aggregated settings/hardware accel info for easy reference
+
 ### Fixed
 - Fix scaling anamorphic content from non-local libraries
 - Fix direct streaming content from Jellyfin that has external subtitles
@@ -11,6 +14,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - Fix Jellyfin, Emby and Plex library scans that wouldn't work in certain timezones
 - Fix song normalization to match FFmpeg Profile bit depth
 - Fix bug playing some external subtitle files (e.g. with an apostrophe in the file name)
+- Fix bug detecting VAAPI capabilities when no device is selected in active FFmpeg Profile
 
 ### Changed
 - Ignore case of video and audio file extensions in local folder scanner